Minimize upstream diff ahead of the heapless merge
A multi-agent review of the try_const_num_traits→heapless delta against
RustCrypto/RSA master flagged divergence that isn't required by the
fork's functionality. This applies the fixes:

- Restore LICENSE-APACHE, LICENSE-MIT, release.toml to upstream bytes
  and re-add ci.yml's upstream trailing space — all four were silently
  mutated by the pre-commit whitespace fixers. Widen the pre-commit
  `exclude` regex (LICENSE-*, release.toml, .github/) so
  upstream-tracked files stay byte-identical going forward.
- rsa_decrypt / rsa_decrypt_and_check: revert to upstream's
  `&impl PrivateKeyParts<...>` signatures and method-call syntax; the
  genericization only needs the `<BoxedUint>` type argument, not the
  named-K + UFCS reshape. Also drops the forked test turbofish in
  key.rs. The primes-empty guard in the CRT arm stays (deliberate
  hardening for keys with CRT accessors but no stored primes).
- Non-CRT decrypt arm: restore upstream's pow_mod_params call
  (rsa_private_op reroute was behavior-neutral churn).
- key.rs: spell the four boxed-only impls (Eq/PartialEq/AsRef/Hash)
  with the RsaPrivateKey/RsaPublicKey aliases, matching upstream and
  the Serialize impls in the same file.
- Restore `alloc` (not `private-key`) cfg on left_pad and the
  algorithms::{pad,pkcs1v15} test modules — the narrowing silently
  dropped them from alloc-without-private-key builds.
- Move the pss blinded-sign round-trip test out of upstream-tracked
  pss/signing_key.rs into fork-only pss/generic_signing_key.rs, which
  owns the method under test.
- footprint/cortex-m: fixed-bigint 0.4.0 → 0.4.1, matching the crate
  manifest.
